Fix fileSave return type
The docs for fileSave() are correct, the return type is optional.
However, the typing promised to always return a defined
FileSystemHandle. This patch makes the return type optional to forgive
the legacy behavior.

`null` is used over `undefined` to match project typing conventions. As
such, a functional change to unconditionally `return null` was necessary
in the legacy implementation.